+++ src/scenes/menu/menu.module.scss
@@ -3,24 +3,22 @@
 .container {
   position: fixed;
   top: 0;
-  z-index: 2;
+  z-index: 4;
   width: 100%;
   height: 80px;
-  transition: height 0.5s;
   background-color: $color-secondary;
   box-shadow: -1px 3px 3px -3px $color-tertiary;
   -webkit-box-shadow: -1px 3px 3px -3px $color-tertiary;
   -moz-box-shadow: -1px 3px 3px -3px $color-tertiary;
 }
 
-.menuIsOpen {
-  height: 425px;
-}
 
 .innerContainer {
   max-width: $appWidth;
+  position: relative;
   margin: 0 auto;
   display: flex;
   justify-content: space-between;
-  align-items: flex-start;
+  align-items: center;
+  z-index: 3;
 }
