+++ src/scenes/menu/menu.tsx
@@ -2,22 +2,20 @@ import React, { useState } from "react";
 import styles from "./menu.module.scss";
 import { Title } from "./title";
 import { MenuDesktop } from "./menuDesktop";
-import { MenuMobile } from "./menuMobile";
-import classNames from "classnames";
+import { MenuMobileIcon } from "./menuMobileIcon";
+import { MenuMobileList } from "./menuMobileList";
 
 export function Menu() {
   const [menuIsOpen, setMenuIsOpen] = useState(false);
 
-  const containerClass = classNames(styles.container, {
-    [styles.menuIsOpen]: menuIsOpen,
-  });
   return (
-    <div className={containerClass}>
+    <div className={styles.container}>
       <div className={styles.innerContainer}>
         <Title />
-        <MenuMobile menuIsOpen={menuIsOpen} setMenuIsOpen={setMenuIsOpen} />
+        <MenuMobileIcon menuIsOpen={menuIsOpen} setMenuIsOpen={setMenuIsOpen} />
         <MenuDesktop />
       </div>
+      <MenuMobileList menuIsOpen={menuIsOpen} setMenuIsOpen={setMenuIsOpen} />
     </div>
   );
 }
