+++ src/scenes/menu/menuList.module.scss
@@ -1,0 +1,3 @@
+.list{
+    margin: 0;
+}
