--- src/scenes/menu/menuMobile.module.scss
@@ -1,12 +1,0 @@
-@import "../../variables.scss";
-
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-}
-@media screen and (min-width: $laptop) {
-  .container {
-    display: none;
-  }
-}
