+++ src/scenes/menu/menuMobileList.tsx
@@ -1,5 +1,7 @@
 import React from "react";
 import { MenuList } from "./menuList";
+import styles from "./menuMobileList.module.scss";
+import classNames from "classnames";
 
 interface MenuMobileListProps {
   menuIsOpen: boolean;
@@ -10,13 +12,16 @@ export function MenuMobileList({
   menuIsOpen,
   setMenuIsOpen,
 }: MenuMobileListProps) {
+  const handleClick = () => {
+    setMenuIsOpen(false);
+  };
+  const containerClass = classNames(styles.container, {
+    [styles.menuIsOpen]: menuIsOpen,
+  });
+
   return (
-    <>
-      {menuIsOpen && (
-        <div>
-          <MenuList />
-        </div>
-      )}
-    </>
+    <div onClick={handleClick} className={containerClass}>
+      <MenuList />
+    </div>
   );
 }
